requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
django-ninja==1.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
